+++ changelog.md
@@ -8,6 +8,7 @@ Features
 * Stream input from STDIN to consume less memory, adding `--noninteractive` and `--format=` CLI arguments.
 * Remove suggested quoting on completions for identifiers with uppercase.
 * Allow table names to be completed with leading schema names.
+* Soft deprecate the built-in SSH features.
 
 
 Bug Fixes